+++ lib/repository/location_repository.dart
@@ -1,6 +1,23 @@
 import 'package:geolocator/geolocator.dart';
+import 'package:weather_app_flutter_mvvm/datastore/location_datastore_interface.dart';
+import 'package:weather_app_flutter_mvvm/repository/location_repository_interface.dart';
 
-abstract class LocationRepositoryInterface {
-  Future<bool> requestPermission();
-  Future<Position> getCurrentLocation();
+class LocationRepository implements LocationRepositoryInterface {
+  final LocationDataStoreInterface dataStore;
+
+  LocationRepository({required this.dataStore});
+
+  @override
+  Future<bool> requestPermission() async {
+    return await dataStore.requestPermission();
+  }
+
+  @override
+  Future<Position> getCurrentLocation() async {
+    try {
+      return await dataStore.getCurrentLocation();
+    } catch (exception) {
+      throw Exception(exception);
+    }
+  }
 }
